36_2026_1_28_rag_aiagent_blogwriter/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
tavily-python
python-dotenv
